No need to fill the modeline with minor mode lighters
When we diminish a mode, we are saying we want it to continue doing its work for
us, but we no longer want to be reminded of it. It becomes a night worker, like
a janitor; it becomes an invisible man; it remains a component, perhaps an
important one, sometimes an indispensable one, of the mechanism that maintains
the day-people's world, but its place in their thoughts is diminished, usually
to nothing. As we grow old we diminish more and more such thoughts, such people,
usually to nothing. -- Will Mengarini