Add Teodor function to auto-generate bb.edn content
The autogenerated bb.edn refers to deps.edn in the same folder.

This is what you want when you write Babashka-compatible libraries: Launching a
JVM REPL and a Babashka REPL will now give the same libraries.

This is *not* what you want in application code. It would force Babashka to load
all your JVM deps, probably slow, and possibly crashing the REPL before it can
start.